requirements.txt
numpy
torch
vispy
PyYAML
tk
